Restructure as a proper Python package.
This will make it possible to share code more conveniently.  It also
gives us a way to declare the non-standard dependencies of the tools
(currently: pyasn, genson).

The downside is that you have to run the top-level "prepare.sh" script
before the commands work, but at least it's a single concrete thing
now, instead of an unspecified number of undocumented "pip install"
operations.

I picked Flit as the "packaging back end" because it's the most
minimal and least opinionated tool I could find for the job.
I have a lot of opinions about software packaging myself, many
of which are, er, heterodox, so something like Poetry gets in
my way more than it helps.