Prepare for mechanical code reformatting using YAPF.
Add a dev dependency on YAPF to pyproject.toml, and a configuration
pattern; tweak the code in a bunch of places so the mechanical
reformat (in the next commit) doesn't mangle data, argparse setup
chains, and a few other places where consistency across multiple
expressions is desirable.

YAPF, not Black, because I violently disagree with certain choices
made by the Black maintainers; YAPF has better defaults, and allows
you to change the subset of its defaults that are bad.